+++ src/main/java/io/zipcoder/casino/cardclasses/CardGame.java
@@ -1,6 +1,6 @@
 package io.zipcoder.casino.cardclasses;
 
-import io.zipcoder.casino.utilities.Player;
+import io.zipcoder.casino.Player;
 
 import java.util.ArrayList;
