+++ src/test/java/io/zipcoder/casino/PlayerTest.java
@@ -1,6 +1,5 @@
 package io.zipcoder.casino;
 
-import io.zipcoder.casino.utilities.Player;
 import org.junit.Assert;
 import org.junit.Test;
