+++ src/test/java/io/zipcoder/casino/cardclasses/CardGameTest.java
@@ -1,6 +1,6 @@
 package io.zipcoder.casino.cardclasses;
 
-import io.zipcoder.casino.utilities.Player;
+import io.zipcoder.casino.Player;
 import org.junit.Assert;
 import org.junit.Test;
